requirements.txt
python-dotenv
trio
rich
readchar
openai
